bridge/pyproject.toml
[tool.poetry]
name = "planktoscope-mqtt-serial-bridge"
version = "0.1.0"
description = "A bridge which watches for MQTT traffic over the PlanktoScope API and sends messages to a serial device"
authors = ["Ethan Li <lietk12@gmail.com>"]
license = "Apache-2.0 OR BlueOak-1.0.0"
readme = "README.md"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "pypi"
priority = "supplemental"

[[tool.poetry.source]]
name = "piwheels"
url = "https://www.piwheels.org/simple/"
priority = "primary"

[tool.poetry.dependencies]
python = "^3.9"
paho-mqtt = [
  { version = "~2.0.0", source = "pypi", markers = "platform_machine != 'armv7l'" },
  { version = "~2.0.0", source = "piwheels", markers = "platform_machine == 'armv7l'" },
]
pyserial = [
  { version = "~3.5", source = "pypi", markers = "platform_machine != 'armv7l'" },
  { version = "~3.5", source = "piwheels", markers = "platform_machine == 'armv7l'" },
]
